simplify function args
